Remove uses-library for checking Chrome presence.
Chrome does not export library anymore.  Now we check, the presence of
Chrome at runtime and show an alert if it is not installed.  It means that
we can install this app without having Chrome installed first, but it seems
that Android Package Manager does not provide any dependancy handling tools
other than for libraries.